Clean up Vector Normalization tests slightly
